+++ src/underlay/singlehostunderlay/udpoutscheduler.cc
@@ -114,7 +114,7 @@ int UdpOutScheduler::initializeNetwork()
     }
 
     addr.sin_port = htons(overlay->gate("appIn")->getNextGate()->
-                          getOwnerModule()->par("localPort").longValue());
+                          getOwnerModule()->par("localPort").intValue());
 
     cModule* underlayConfigurator =
         (*getSimulation()).getModuleByPath("SingleHostUnderlayNetwork.underlayConfigurator");
